Moved get request to network manager
